Revert "Updated fillapix files based on new analysis."
